Organize project files into a standard React structure
- Created public/ and src/ directories with subdirectories for components, services, hooks, and utils.
- Moved existing JavaScript and JSX files into the new structure.
- Updated import paths within the moved files to reflect their new locations.
- Added missing React/PropTypes imports where necessary.
- Identified critical missing files, including public/index.html and several application components and services.
- Noted the backend-like nature of fcmTwilioNotifier.js and its unlisted dependencies.
