--- index.js
@@ -1,14 +1,0 @@
-import App from './App';
-
-const root = ReactDOM.createRoot(document.getElementById('root'));
-
-root.render(
-  <React.StrictMode>
-    <BrowserRouter>
-      <App />
-    </BrowserRouter>
-  </React.StrictMode>
-);
-
-serviceWorkerRegistration.register();
-reportWebVitals();
