+++ src/index.js
@@ -1,0 +1,19 @@
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import { BrowserRouter } from 'react-router-dom';
+import App from './App';
+// import * as serviceWorkerRegistration from './serviceWorkerRegistration'; // Assuming this file will be created later
+// import reportWebVitals from './reportWebVitals'; // Assuming this file will be created later
+
+const root = ReactDOM.createRoot(document.getElementById('root'));
+
+root.render(
+  <React.StrictMode>
+    <BrowserRouter>
+      <App />
+    </BrowserRouter>
+  </React.StrictMode>
+);
+
+serviceWorkerRegistration.register();
+reportWebVitals();
