+++ src/services/userModerationSubscriptionAnalyticsApi.js
@@ -1,3 +1,5 @@
+import axios from 'axios';
+
 const apiClient = axios.create({
   baseURL: process.env.REACT_APP_API_URL,
   timeout: 10000,
