+++ src/utils/encryptedSocketIoClient.js
